separate Chef filter and search components
